Cleanup Gemfile
The :development group is pointless because Gemfile is relevant for
those who run ruby/openssl tests in this repository only, and not
included in the gem package. Let's list the test dependency in the top
level.

The rdoc library is now a bundled gem in Ruby master, not just in
Fedora's Ruby distributions.