Rename the `SSLContext#ecdh_curves=` to `#groups=`.
Rename the `SSLContext#ecdh_curves=` to `#groups=`, keeping the `#ecdh_curves=`
as an alias of the `#groups=` for backward compatibility.

In OpenSSL version 1.1.1 `SSL_CTX_set1_curves_list` was renamed to
`SSL_CTX_set1_groups_list`, keeping the `SSL_CTX_set1_curves_list` at the
following commit. Because TLS 1.3 renamed the extension.
https://github.com/openssl/openssl/commit/de4d764e3271ce09d28c0d6d7bce3dc9d8b85ab9

As we only support OpenSSL version 1.1.1 or later versions, it is no problem to
use the `SSL_CTX_set1_groups_list`.

For the document of the `ossl_sslctx_set_groups`,
I referred to the following RFC 8446 TLS 1.3 Supported Groups section.
https://datatracker.ietf.org/doc/html/rfc8446#section-4.2.7