Revert "pkey: stop retrying after non-retryable error from OSSL_DECODER"
This reverts commit 5347880c6eb0dd1144de6ee0b1f7a9cf3ee899b8 and
985ba27d63398e2d26bc1ea3d71bdb57550b9045.

These commits attempted to stop processing after the first relevant PEM
block, whether it is successful or not, when the input contains multiple
keys.

It turned out that it cannot be reliably determined using the
OSSL_DECODER API. There is an edge case where OSSL_DECODER_from_bio()
reports "unsupported" even though the input actually contains an error:

https://redirect.github.com/ruby/openssl/pull/931#discussion_r2347813807

Revert the changes for now and keep the existing behavior, as partial
support does not seem worth the added complexity.