CI: Add GitHub Actions ppc64le/s390x cases
This commit adds the GitHub Actions ppc64le/s390x cases using the CI services
provided by the project <https://github.com/IBM/actionspz>.

As a context, we submitted the request to use the CI by the ticket
<https://github.com/IBM/actionspz/issues/4>, and the ruby/openssl is one of the
ruby/* repositories that we want to add the CI cases.

This CI doesn't work in fork repositories. Therefore we create the new job
`test-ibm` to use the `if` syntax to run the `test-ibm` job only in ruby/openssl
repository.

I am using YAML's anchors and aliases feature to avoid duplicated logic between
the `test` and `test-ibm` jobs. The document is below.
https://docs.github.com/en/actions/reference/workflows-and-actions/reusing-workflow-configurations#yaml-anchors-and-aliases