ssl: avoid subclassing String
Before Ruby 3.0, String#slice! on a String subclass returned a new
string with the same class as the receiver. As a result, a few methods
on OpenSSL::Buffering leak the internal class Buffer.

Replace Buffer with refinements that are enabled only on Ruby < 3.4,
where String#append_as_bytes is absent.